requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.65.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
scikit-image>=0.21.0
scipy>=1.10.0
noise>=1.2.2
imgaug>=0.4.0
tensorboard>=2.13.0
scikit-learn>=1.3.0
pandas>=2.0.0